requirements.txt
# === Основа FastAPI ===
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9

# === JWT и Безопасность ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0

# === Email ===
fastapi-mail==1.4.1
jinja2==3.1.4

# === База данных ===
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# === Конфигурация ===
pydantic-settings==2.5.2
python-dotenv==1.0.1

# === Rate Limiting ===
slowapi==0.1.9

# === Тестирование ===
pytest==8.3.3
httpx==0.27.2

# === Production (опционально) ===
# gunicorn==23.0.0
# celery==5.4.0
# redis==5.1.1
# loguru==0.7.2
# sentry-sdk[fastapi]==2.14.0
